Enable nestStandalone for Zendesk
