+++ packages/adapter-components/src/filters/referenced_instance_names.ts
@@ -126,6 +126,15 @@ const getInstanceNameDependencies = (
   return referencedInstances
 }
 
+const isStandalone = (instance: InstanceElement, configByType: Record<string, TransformationConfig>): boolean => {
+  const parentElemID = getFirstParentElemId(instance)
+  if (parentElemID === undefined) {
+    return false
+  }
+  return configByType[parentElemID.typeName]?.standaloneFields?.map(field => field.fieldName)
+    .includes(instance.elemID.name) ?? false
+}
+
 /* Calculates the new instance name and file path */
 const createInstanceNameAndFilePath = (
   instance: InstanceElement,
@@ -138,7 +147,7 @@ const createInstanceNameAndFilePath = (
   const newName = joinInstanceNameParts(newNameParts) ?? instance.elemID.name
   const parentName = idConfig.extendsParentId ? getFirstParentElemId(instance)?.name : undefined
   const { typeName, adapter } = instance.elemID
-  const { fileNameFields, serviceIdField } = configByType[typeName]
+  const { fileNameFields, serviceIdField, nestStandaloneInstances } = configByType[typeName]
 
   const newNaclName = getInstanceNaclName({
     entry: instance.value,
@@ -158,6 +167,8 @@ const createInstanceNameAndFilePath = (
     isSettingType: configByType[typeName].isSingleton ?? false,
     nameMapping: configByType[typeName].nameMapping,
     adapterName: adapter,
+    nestedPaths: nestStandaloneInstances && isStandalone(instance, configByType)
+      ? [...instance.path?.slice(2, instance.path?.length - 1) ?? []] : undefined,
   })
   return { newNaclName, filePath }
 }
@@ -280,7 +291,7 @@ const shouldChangeElemId = (idFields: string[], extendsParentId: boolean | undef
 /* Create a graph with instance names as nodes and instance name dependencies as edges */
 const createGraph = (
   instances: InstanceElement[],
-  instanceToIdConfig: { instance: InstanceElement; idConfig: TransformationIdConfig }[],
+  instanceToIdConfig: { instance: InstanceElement; idConfig: TransformationIdConfig }[]
 ): DAG<InstanceElement> => {
   const duplicateElemIds = new Set(findDuplicates(instances.map(i => i.elemID.getFullName())))
   const duplicateIdsToLog = new Set<string>()
@@ -317,7 +328,8 @@ export const createReferenceIndex = (
   allInstances: InstanceElement[],
   instancesNamesToRename: Set<string>,
 ): Record<string, { path: ElemID; value: ReferenceExpression }[]> => {
-  const allReferences = allInstances.flatMap(instance => getReferencesToElemIds(instance, instancesNamesToRename))
+  const allReferences = allInstances
+    .flatMap(instance => getReferencesToElemIds(instance, instancesNamesToRename))
   const referenceIndex = _(allReferences)
     .groupBy(({ value }) => value.elemID.createTopLevelParentID().parent.getFullName())
     .value()
@@ -393,17 +405,17 @@ export const referencedInstanceNamesFilterCreator: <
   TClient,
   TContext extends { apiDefinitions: AdapterApiConfig },
   TResult extends void | filter.FilterResult = void,
->(
+  >(
   customApiDefinitions?: AdapterApiConfig,
 ) => FilterCreator<TClient, TContext, TResult> =
   customApiDefinitions =>
-  ({ config, getElemIdFunc }) => ({
-    name: 'referencedInstanceNames',
-    onFetch: async (elements: Element[]) => {
-      const apiDefinitions = customApiDefinitions ?? config.apiDefinitions
-      const transformationDefault = apiDefinitions.typeDefaults.transformation
-      const configByType = apiDefinitions.types
-      const transformationByType = getTransformationConfigByType(configByType)
-      await addReferencesToInstanceNames(elements, transformationByType, transformationDefault, getElemIdFunc)
-    },
-  })
+    ({ config, getElemIdFunc }) => ({
+      name: 'referencedInstanceNames',
+      onFetch: async (elements: Element[]) => {
+        const apiDefinitions = customApiDefinitions ?? config.apiDefinitions
+        const transformationDefault = apiDefinitions.typeDefaults.transformation
+        const configByType = apiDefinitions.types
+        const transformationByType = getTransformationConfigByType(configByType)
+        await addReferencesToInstanceNames(elements, transformationByType, transformationDefault, getElemIdFunc)
+      },
+    })
